Added insert of dummy user to sql script
In the new database schema, every recipe belongs to a given user. Because webkochbuch does not support
user management completely at the moment, a simple dummy user is created with the create_tables sql
script. This dummy user will implicitly be set as the owner of newly created recipes.

JIRA:
CG-78